Insert acquire fences into 3D pipeline as needed.
When we fallback to 3D composition, we need to wait
for acquire fences to be signalled before reading from
the buffer. Instead of polling let's insert fence into the
3D queue.

Jira: None.
Test: Reduced flickering when launching applications

+++ common/compositor/compositor.h
@@ -50,7 +50,7 @@ class Compositor {
                      NativeBufferHandler *buffer_handler, uint32_t width,
                      uint32_t height, HWCNativeHandle output_handle,
                      int32_t *retire_fence);
-  void InsertFence(uint64_t fence);
+  void InsertFence(int32_t fence);
 
   Renderer *GetRenderer() const {
     return renderer_.get();
